+++ changelog.md
@@ -11,6 +11,7 @@
 * `BaseService` now contains a `MetadataCollection` to tightly couple the metadata to the associated service.
 * Added `Services`, a new class which contains a copy of each `BaseService` supported by the SDK.
 * Added a new connection method `connect_with_token` which allows you to connect to the EWB using an access token.
+* Added a new connection method `connect_with_token_insecure` which allows you to make an insecure connection to the EWB using an access token without TLS.
 
 ### Enhancements
 * None.